+++ kite-core/src/main/kotlin/com/tang/kite/config/DataSourceConfig.kt
@@ -1,0 +1,16 @@
+package com.tang.kite.config
+
+/**
+ * Datasource properties for datasource configuration.
+ *
+ * @author Tang
+ */
+object DataSourceConfig {
+
+    /**
+     * Whether to override the existing data source when registering a new one.
+     */
+    @JvmStatic
+    var override: Boolean = false
+
+}